+++ test/ConverterHelper.scala
@@ -1,0 +1,12 @@
+import org.streum.configrity.converter._
+
+
+object ConverterHelper {
+  
+  def convert[A]( opt: Option[String] )(implicit converter: ValueConverter[A] )  =
+    converter(opt)
+
+  def convert[A]( opt: String )(implicit converter: ValueConverter[A] )  =
+    converter.parse(opt)
+
+}